--- src/Utility/connected_graph.txt
@@ -1,1 +1,0 @@
-{"1":["2","3"],"2":["1","3"], "3":["1", "2", "4" ], "4":["3"]}
